LCManager moved to impl and Changed the LifeCycle Reading from registry to Database
